Refactor. Changed from AlphaVantage to Coin Market Cap #59
