+++ src/test/java/nitezh/ministock/dataaccess/CoinMarketCapRepoTest.java
@@ -3,8 +3,12 @@ package nitezh.ministock.dataaccess;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+import org.junit.Before;
 import org.junit.Test;
 
+import java.util.HashMap;
+
+import nitezh.ministock.domain.StockQuote;
 import nitezh.ministock.mocks.MockCache;
 
 import static junit.framework.Assert.assertEquals;
@@ -20,12 +24,16 @@ import static org.junit.Assert.assertNotNull;
 public class CoinMarketCapRepoTest {
     CoinMarketCapRepo coinMarketCapRepo;
 
+    @Before
+    public void setup(){
+        coinMarketCapRepo = new CoinMarketCapRepo(new FxChangeRepository());
+    }
+
     @Test
     public void getAllDefaultAPIData() throws JSONException{  //By Default, we should be getting the first 100 stocks
         String blankID = "";  //should get all the data if no id is provided
         int NumberOfTotalDefaultStocks = 100;  //Number of total stocks from CoinMarketRepo by default
 
-        coinMarketCapRepo = new CoinMarketCapRepo(new FxChangeRepository());
         JSONArray jsonArray = coinMarketCapRepo.retrieveQuotesAsJson(blankID, 0);
         assertNotNull(jsonArray);
         assertEquals(NumberOfTotalDefaultStocks , jsonArray.length());  //we should be getting every stock
@@ -51,7 +59,6 @@ public class CoinMarketCapRepoTest {
     @Test
     public void getAPIDataByID() throws JSONException{
         String id = "bitcoin"; //will get the data only related to bitcoin
-        coinMarketCapRepo = new CoinMarketCapRepo(new FxChangeRepository());
         JSONArray jsonArray = coinMarketCapRepo.retrieveQuotesAsJson(id, 0);
         assertNotNull(jsonArray);
         assertEquals(1, jsonArray.length());  //we should only be getting bitcoin
@@ -79,7 +86,6 @@ public class CoinMarketCapRepoTest {
         String blankID = "";  //should get all possible data if no id is provided
         int limit = 90;  //Limit the number of stocks to 90
 
-        coinMarketCapRepo = new CoinMarketCapRepo(new FxChangeRepository());
         JSONArray jsonArray = coinMarketCapRepo.retrieveQuotesAsJson(blankID, limit);
         assertNotNull(jsonArray);
 
@@ -100,5 +106,24 @@ public class CoinMarketCapRepoTest {
             assertFalse(jsonObject.optString("percent_change_7d").isEmpty());
             assertFalse(jsonObject.optString("last_updated").isEmpty());
        }
+    }
+
+    @Test
+    public void getQuotes() throws JSONException{
+        // Arrange
+        String symbol ="bitcoin";
+        String market = "USD";
+
+        // Act
+        HashMap<String, StockQuote> stockQuotes = coinMarketCapRepo.getQuotes(symbol);
+
+        // Assert
+        assertEquals(1, stockQuotes.size());
+        StockQuote BTCQuote = stockQuotes.get("BTC");
+        assertEquals("BTC", BTCQuote.getSymbol());
+        assertEquals("Bitcoin", BTCQuote.getName());
+        assertFalse(BTCQuote.getVolume().isEmpty());
+        assertFalse(BTCQuote.getPrice().isEmpty());
+        assertFalse(BTCQuote.getPercent().isEmpty());
     }
 }
